go/ir: always generate wrappers for generics, annotate generic calls with type arguments
We elided generic wrappers for functions that have no generic arguments
or return values, such as 'func[T any]()'. However, this is inaccurate,
as these functions still need to be instantiated, and their behavior may
differ under different type arguments.

We also enrich calls with the concrete types passed as arguments to type
parameters. Currently, these are only used in generic wrappers, but in
the future they would allow inlining wrappers of generics.