go/ast/astutil: don't panic trying to copy FuncType
In gh-1134, I argued in favour of extending CopyExpr to use reflection
to handle all types, under the assumption that we missed many types.
I have now decided against that, for two reasons:

1. There weren't many unhandled types left
2. We cannot copy comments because of limitations in go/ast, and it
seems better to refuse generating suggested fixes than to swallow
comments. This is a general problem, but function literals are
particularly likely to contain comments.

Thus, we just add FuncType to the list of types we cannot copy. The
effect is that several checks in the quickfix package will not generate
suggested fixes in some cases.

Closes gh-1134