lintcmd: make binary output OS agnostic
We were storing absolute paths and full token.Position in the binary
output. This works fine when merging multiple runs from the same OS, but
not so when dealing with multiple OSs, which have different path
separators and different absolute paths. We also have to zero out
token.Position.Offset to handle different newlines on different OSs.