go/ir: improve lifting of allocs by splitting on liftability boundary
Instead of skipping allocs that are used in unliftable ways, split
allocs into two parts: one used in liftable ways and one in unliftable
ways. Uses that are lfitable and aren't reachable by unliftable ones can
be lifted as normal, leaving only a subset of a function that cannot be
lifted. We emit the necessary stores at liftable->unliftable boundaries
to propagate the last lifted values to the unliftable allocs.

This makes the "ineffective field assignment" check less useful, because
it only considers receivers, and after splitting, the value is
technically no longer the receiver. We will address this after
implementing SROA, which will break even more of that check.