unused: reimplement check using the AST
This is a complete rewrite of U1000, using a walk over the AST instead
of the IR. This addresses the many issues where the IR is too clever and
too far removed from the actual code to correctly track usedness of
identifiers.

For example, a value that is only used as part of unsafe.Sizeof will
become a constant in the IR, losing its connection to the identifier of
the value. Another common issue is skipped unit tests, which, due to the
IR's analysis of control flow, look mostly empty. Users, however, don't
expect identifiers to be unused when their only uses are in skipped
tests.

Additionally, we've fixed the way objects in generated files are
ignored. Instead of simply suppressing the output of their diagnostics,
generated objects are now marked reachable from the root, so that their
uses are tracked correctly.

Another change lies in the format of the returned data. Instead of
producing lists, which are enough to handle unexported objects, we now
return the full usedness graph, in a format that allows merging multiple
graphs, to compute the usedness of exported API incrementally. This
isn't currently being used anywhere, but will allow us to reintroduce a
dedicated tool that can find unused exported objects, among other
things. The Options type contains booleans for the many ways in which
the new tool can be configured in.

Finally, we've improved our unit tests to explicitly track identifiers
that were quietened (unused objects that belong to outer, unused
objects), and to verify the names of objects, not just their locations.

Overall, this implementation reduces both the number of false positives
and false negatives.

Closes: gh-507
Closes: gh-633
Closes: gh-810
Closes: gh-812
Closes: gh-1199
Closes: gh-1249
Closes: gh-1282
Closes: gh-1333