knowledge: record io/ioutil as having been deprecated in Go 1.19
Go 1.16 informally deprecated io/ioutil, but only added the markers in
Go 1.19. We recorded it as having been deprecated in 1.16 anyway, based
on the idea that someone running Go 1.19 and using 'staticcheck -go
1.16' should be notified of io/iotuil having been deprecated.

However, because SA1019 relies on the comment marker to know that
something has been deprecated, a user targeting Go 1.16 and using Go
1.18 will not see any warnings, but once they upgrade to Go 1.19 - while
still targeting Go 1.16 - they will suddenly see warnings. This is
confusing to experience and difficult to explain.

We should stick to Go's official word: io/ioutil has been deprecated in
Go 1.19.