Move all test assets in example.com directories
Our heuristic for which packages are in the standard library is to look
for a dot in the import path. The vast majority of our test code was
incorrectly classified as in the standard library. This only affects
some checks (such as the one for deprecated code), but this behavior has
cost us over an hour of debugging.