SA1019: account for all combinations of test packages using each other
A single Go package results in three packages when testing: the package
under test with its internal tests, a package for external tests, and a
package main that runs the tests.

We need to allow external tests to import the package under test, and
the test's main to import both the package under test as well as its
external tests, and to use the objects defined in them.

Closes gh-1285