go/ir: don't check for duplicate function names in tests
Generic functions can instantiate generic types with type parameters,
which can lead to methods with duplicated names. We handle this by using
a generic-specific synthetic constant. However, we also generate
wrappers for these instantiated methods, which leads to more duplicated
names, and the code doesn't easily let us figure out that the wrapper is
for an instantiated method. Instead of working around this further,
remove the test for unique function names. Function names weren't unique
in practice, anyway, because of test variants of packages.