go/ir: handle labeled statements with blank labels
A labeled statement with a blank label can't be the target of a goto,
break, or continue statement, so we don't need to create a new block for
it in builder.stmt.

Function.labelledBlock, which previously would be called to get a new
block, panics for blank labels anyway, because the type-checker doesn't
associate a blank label's *ast.Ident with a *types.Object.

(cherry picked from commit d10b068e4dcb22cb6bf53ab2fd25a347d3642f1f in
x/tools)

This is an alternative to our 0c085726a43c410cac20e948d19b3c63b8c9b379.
We backport this change to avoid unnecessary differences between go/ir
and go/ssa.

Co-authored-by: Dominik Honnef <dominik@honnef.co>