lintcmd: make -check case-insensitive, error on unknown checks
This changes it so that the check is case-insensitive, and so it will
return an error on unknown checks. Previously:

	% staticcheck -checks=sa1000 ./lintcmd
	[no output, exit 0]

	% staticcheck -checks=banana ./lintcmd
	[no output, exit 0]

	% staticcheck -checks=SA1000 ./lintcmd
	lintcmd/cmd.go:270:34: error parsing regexp: missing closing ): `$^(` (SA1000)

This is massively confusing! When I was doing my testing for the other
PRs yesterday I found it odd that I didn't get any error for the defer
one, since I *knew* there was at least one module with this error.
Turned out it just always passed because I used lower-case.

I wouldn't be surprised if tons of people have made this mistake and
have run (or are running) staticcheck runs that don't actually do
anything.

All of this also applies to checks = [..] from the config file.