website: don't check checks.json into git anymore
checks.json is a generated file. We initially committed it so that `hugo
serve` would work without any preceding build steps. However, the file
is almost always out of date because we don't update it when adding new
checks. Furthermore, the required build step is trivial and only
requires Go.