unused: handle structs.HostLayout
The HostLayout marker, added in Go 1.23 (see [1] for its background),
signals to the Go compiler that the layout of a struct -- with regard to
field order and size -- matters, to be compatible with some externally
defined memory layout, such as for interopability with C.

This implies that the presence of all fields is important, and thus we
shouldn't flag any of them as unused. This augments our unsafe.Pointer
conversion logic and catches some edge cases that it misses, as well as
more exotic scenarios that don't involve any use of unsafe.

It is an open question[2] which uses of structs.HostLayout will be
detected by the compiler (wrt named types, embedding, generics, etc). We
implement the strictest sensible interpretation, requiring a field
directly in the struct, using exactly the structs.HostLayout type or an
alias thereof. The field itself may be emebdded, but embedding another
struct with such a field doesn't affect the embedder.

Our HostLayout logic doesn't replace our unsafe.Pointer logic, for
several reasons:

- Virtually nobody is using HostLayout yet.

- It is possible that some uses of unsafe.Pointer will continue to work
  even without the marker and that the compiler will detect them.

[1]: https://golang.org/issues/66408
[2]: https://golang.org/issues/66408#issuecomment-2120644459

Closes: gh-1604