pattern: run lexer in a lightweight coroutine
Previously, the lexer ran as a separate goroutine that fed a
channel with items. This change expresses the lexer as an
iter.Seq[item], and then uses the iter.Pull mechanism to
run it as a lightweight coroutine. This reduces the running
time of the added by about 32%.

$ go test  -benchtime=10s -bench=. ./pattern

 before
 BenchmarkParser-8   	  649060	     19051 ns/op
 after
 BenchmarkParser-8   	  908734	     12971 ns/op

Interestingly, the total time spent by gopls in Parse,
measured by bracketing Parse with time.Now/time.Sub calls,
consistently reduces by a factor of around 20x (down to 800us,
originally 15ms). I can't yet explain this discrepancy.